add dashboard route to analytics routes, fix registrationdates and auth
