+++ api/lib/analytics.js
@@ -1,25 +1,43 @@
 const express = require('express')
 const router = express.Router()
+const moment = require('moment')
 const db = require('./adapters/db')
 
-const reqToken = (req, res, next) => {
-  const token = req.headers['Authorization']
+const reqToken = (req, _, next) => {
+  const token = req.headers.authorization
   if (token !== process.env.DASHBOARD_AUTH_TOKEN) {
     throw new Error('not authed')
   }
   next()
 }
 
-router.get('/numUsers', reqToken, (req, res) => {
-  res.json(db.getAllUsers().length)
-})
-
-router.get('/totalSteps', reqToken, (req, res) => {
+router.get('/totalSteps', reqToken, (_, res) => {
   res.json(db.getTotalSteps())
 })
 
-router.get('/userRegistrations', reqToken, (req, res) => {
-  res.json(db.getAllUsers().map(({ created }) => created))
+router.get('/userRegistrations', reqToken, async (_, res) => {
+  const users = await db.getAllUsers()
+  res.json(users.filter((u) => u.created).map(({ created }) => created))
+})
+
+router.get('/dashboard', async (_, res) => {
+  const today = moment().startOf('day').toDate()
+  const oneWeekAgo = moment().subtract(7, 'days').startOf('day').toDate()
+
+  res.send({
+    users: await db.userCount({}),
+    usersToday: await db.userCount({ from: today }),
+    usersLastSevenDays: await db.userCount({ from: oneWeekAgo }),
+    sessions: await db.analyticsCount({ event: 'openApp' }),
+    sessionsToday: await db.analyticsCount({
+      event: 'openApp',
+      from: today,
+    }),
+    sessionsLastSevenDays: await db.analyticsCount({
+      event: 'openApp',
+      from: oneWeekAgo,
+    }),
+  })
 })
 
 module.exports = router
