+++ api/lib/index.js
@@ -160,25 +160,6 @@ app.post('/analytics', async (req, res) => {
 
   res.sendStatus(200)
 })
-app.get('/dashboard', async (req, res) => {
-  const today = moment().startOf('day').toDate()
-  const oneWeekAgo = moment().subtract(7, 'days').startOf('day').toDate()
-
-  res.send({
-    users: await db.userCount({}),
-    usersToday: await db.userCount({ from: today }),
-    usersLastSevenDays: await db.userCount({ from: oneWeekAgo }),
-    sessions: await db.analyticsCount({ event: 'openApp' }),
-    sessionsToday: await db.analyticsCount({
-      event: 'openApp',
-      from: today,
-    }),
-    sessionsLastSevenDays: await db.analyticsCount({
-      event: 'openApp',
-      from: oneWeekAgo,
-    }),
-  })
-})
 
 app.post('/ping', async (_, res) => res.send({ ok: true }))
